Rename prediction model to avoid conflict with other implementation
