+++ modules/world/predictor/commons.hpp
