+++ modules/world/predictor/prediction.cpp
@@ -1,4 +1,4 @@
-#include "modules/world/prediction/prediction.hpp"
+#include "modules/world/predictor/prediction.hpp"
 
 namespace modules {
 namespace world {
